+++ app/main.py
@@ -6,13 +6,16 @@ from app.db import crud
 from app.db import models
 from app.db import schemas
 from app.db.database import SessionLocal, engine
-import app.routes.auth as auth
 from logs.logging_config import setup_logging
+
+import app.routes.auth as authRouter
+import app.routes.user as userRouter
 
 models.Base.metadata.create_all(bind=engine)
 
 app = FastAPI()
-app.include_router(auth.router)
+app.include_router(authRouter.router)
+app.include_router(userRouter.router)
 
 @app.on_event("startup")
 def startup_event():
@@ -42,20 +45,7 @@ def get_db():
 db_dependency = Annotated[Session, Depends(get_db)]
 
 # used to make sure user is authenticated when accessing routes
-user_dependency = Annotated[dict, Depends(auth.get_current_user)]
-
-@app.get("/users/", response_model=list[schemas.User])
-async def read_users(db: db_dependency, skip: int = 0, limit: int = 100):
-    users = crud.get_users(db, skip=skip, limit=limit)
-    return users
-
-
-@app.get("/users/{user_id}", response_model=schemas.User)
-async def read_user(user_id: int, db: db_dependency):
-    db_user = crud.get_user(db, user_id=user_id)
-    if db_user is None:
-        raise HTTPException(status_code=404, detail="User not found")
-    return db_user
+user_dependency = Annotated[dict, Depends(authRouter.get_current_user)]
 
 @app.get("/profiles/{profile_id}", response_model=schemas.Profile)
 async def read_profile(profile_id: int, db: db_dependency):
